Fix integration tests when system `sh` is recent bash
My main machine is now a Fedora box. OS X ships with an ancient
version of bash installed as `sh` due to the great GPL purge[0],
so my `sh` is now much more recent.

This broke some tests which were relying on some non-portable
details of the echo builtin. Switched to using printf, which
should be more consistent across platforms.

[0] http://meta.ath0.com/2012/02/05/apples-great-gpl-purge/