Fix integration tests with dash
Integration tests run with bash, dash, and whatever's installed as
`sh`, to ensure compatibility with a wide range of systems.

This commit to dash:

https://git.kernel.org/pub/scm/utils/dash/dash.git/commit/?id=6900ff60ef7347a8c1445853a8f4689808e0976e

Changed the way that the shell escapes special characters. This commit
modifies our tests to be compatible with dash before and after the
changes, and should fix the Travis build.