Reform positional argument parsing
This diff makes positional argument parsing much cleaner, along with
adding a bunch of tests. Just's positional argument parsing is rather,
complex, so hopefully this reform allows it to both be correct and stay
correct.

As part of the reform, `just ..` will now work, and has the same effect
as `just ../`. `just .` is also accepted, but is a noop.

As part of this change, it is now an error to pass arguments or
overrides to subcommands that do not accept them, namely `--dump`,
`--edit`, `--list`, `--show`, and `--summary`. It is also an error to
pass arguments to `--evaluate`, although `--evaluate` does of course
still accept overrides.

The above is a breaking change, but hopefully one which is worth it, as it
will allow us to add arguments to subcommands which did not previously
take them, if we so desire.