Reform indentation handling
Improve indentation handling in preparation for implementing inline
submodules. This changes the lexer to only parse freeform text inside
the first indent after a ':', so that just can be extended with new
indented contstructs which are not recipe bodies. In addition, the lexer
should now handle mutiple levels of indentation correctly.