Document how to run recipes after a recipe is done
Add a section to the readme documenting how to shell out to Just at the
end of a recipe, to simulate dependencies that run after a recipe,
instead of before.

This is certainly not as good as having an explicit syntax for this, but
it's a common question so it would be good to document the current best
workaround.