just step one - disable the DuplicateRecipe Error
Change workflow so that subsequent recipes with the same
name over-ride previously defined recipes.  In effect there
is no longer a concept of a duplicate.  It is now a redefinition.

Will need to do the same thing to duplicate alias workflow.

The last step will be to define a new keyword: import.  The
import keyword will specify from which file to import recipes,
aliases and variables.