import added --allow-duplicates to CLI flags
Used --dry-run as a pattern to add --allow-duplicates to the CLI
flags configuration.  Updated the expected help asset to pass its
test.

The business logic around the ability to allow duplications is still
TBD.  This just puts the flag in place and connects it true/false value
to the global config object.  The default value is false.