import re-enabled the duplicate error message
As a test I first disabled the duplicate recipe error messages and
tests.  This commit re-enables those errors.

Working in the analyze_recipe function to add the check to see
if all0w_duplicates is true HOWEVER, in the context of the function
the "config" object is not available.  I thought it was global
but nooooo.  Need some kind of RUST magic to make it available
in the function and I don't do RUST magic.