Box compile error
Running clippy against just, the vast majority of the errors are:

```
error: the `Err`-variant returned from this function is very large
 --> src/unresolved_recipe.rs:9:8
  |
9 |   ) -> CompileResult<'src, Recipe<'src>> {
  |        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ the `Err`-variant is at least 144 bytes
  |
  = help: try reducing the size of `compile_error::CompileError<'src>`, for example by boxing large elements or replacing it with `Box<compile_error::CompileError<'src>>`
  = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#result_large_err
```

The large size of `CompileResult` is mostly caused by `CompileErrorKind`
being a large type within `CompileError`. So, Box-ing that type on
`CompileError` solves the `CompileResult` problem without requiring too
many additional code changes.

Most of the time, just will only be compiling one value of type
`CompileResult`, which will hopefully most of the time not even have the
error variant on it. So Boxing (part of) that variant in order to reduce
the size of `CompileResult`, which will get passed between many
different functions, makes sense.