Fix windows support in www/install.sh
The current version of install.sh does not quite fully support
installing windows binaries.

This change fixes that by:

 - handling the fact that windows releases are zips not tarballs
 - handling `uname -s` output being funky on MINGW compiled bash

Additionally, it adds a step to the CI tests to test install.sh works on
multiple platforms. These tests uncovered that `need rev` was failing on
windows bash.  However, `rev` is not actually used any more, so
I removed that and `need basename`, as that doesn't seem to be used any
more either.