Use Compilation type
Compiler::compile() now outputs a struct `Compilation` instead of a
tuple of `Ast` and `Justfile`. Right now this `Compilation` struct
contains the same ast and justfile, but abstracting this output into a
new type will make it easier to handle justfiles with includes, which
will necessarily have multiple Ast's (from different included files) as
part of a coherent structure.