feat: add `semver_matches` utility function
While coordinating different programs using `just`, it's often
necessary to compare the *versions* of different binaries or
utilitiies. Assuming that the utility in question providers a
`--version` switch and can reasonably print out a version, being able
to match semver of the utility to an expectation is very useful.

This commit adds a `semver_matches` utility which utilizes the
`semver` crate (https://crates.io/crates/semver) to provide matching
functionality to `Justfile`s.

Signed-off-by: vados <vados@vadosware.io>
Co-authored-by: Casey Rodarmor <casey@rodarmor.com>