Chooser: Pass "--unstable" to the preview command
Pass "--unstable" to the preview command for the chooser, so that
recipe previews can work even when the justfile relies on unstable
features, such as "!include" directives.

Note that we're not taking the user by surprise by enabling unstable
features in the preview: when the justfile includes unstable features,
the user will need to pass "--unstable" explicitely to call "just
--choose --unstable" already, or "just" will exit before launching the
chooser.