Chooser: Pass justfile path to preview command for default chooser
The default chooser is fzf, and we pass it a command to generate the
preview with "just --show" from the recipe name. This has been working
well as long as we rely on the default justfile available in the
directory, given that the preview command can find it too. However,
passing "--chose" alongside "--justfile" to select a specific file
results in the preview command not being able to find the right justfile
to process.

To address this issue, we turn the const string defining the preview
command into a function that takes the internal representation of the
justfile as an argument, and updates the preview command with the right
file.

Fixes: 5f9ac39b038c ("Use `just --show` in default chooser (#1539)")
Suggested-by: Casey Rodarmor <casey@rodarmor.com>