Fix recipe doc alignment
`recipe.private` doesn't take into account recipes with the `[private]`
attribute, so private recipes affect the alignment calculation of the
the recipe docs. This causes them to align too far to the right.

Use `recipe.public()` instead which appears to be the canonical way of
differentiating between public and private.