1748-support-multiple-dotenv-files
Dearest Reviewer,

I use more then dotenv for different reasons. I saw the open issue and took
a stab at it. This allow for more then one set value but only for the one
type.

It was suggested in the issue to support an array syntax ["file", "file2"].
This is not supported today and I didnt look in to changing the parser to support that.

I did consider supporting a csv value like "file,file2" but I didnt move forward on that.

Some things to note:
It does expect the set lines to be parsed in order. I am unsure if that is true today.
It could load files from different subpaths. This could be changed to only load from the path
where the first file is found.

I do like that one file or 10 files they all use the same path.

I like using just. Thank you for your work.
Becker

N.B. The test functions::env_var_functions failed for me on master and my branch.