Changes the env() function so it returns justfile environment variables
Currently, if a variable is exported in a Justfile, `env` (also `env_var` and `env_var_or_default`), will not return it. This commit adds a check of the contexts scope for variables there.

Additionally, this commit merges the functionality of `env`, `env_var` and `env_var_or_default` under `env`, with the others calling env, to reduce the duplication that was present.