Add `metadata` attribute for adding custom metadata to recipes
This PR adds `metadata` attribute, which can appear multiple times on a recipe.
The attribute takes a list of strings and its value can be inspected via
JSON dump of the `justfile`.

The primary motivation is custom tooling built around `just` that can be
used to, for example, automate creation of CI pipelines and other
pre-processing that is outside of `just`'s scope.

Fixes: https://github.com/casey/just/issues/2703