Print executed commands if run in verbose mode
It is currently difficult to debug what went wrong in cases such as OS
exec errors, because verbose output prints the file but does not say how
it is run. Update spawning to print executed commands with loquacious
and above verbosity. The result when run with `-v` is something like the
following:

    + cd "/home/user/workspace" && "/run/user/1000/just/just-kynaKD/configure"

The `+` syntax is meant to match shell scripts run with `-x`. This gets
coloring to differentiate it.

Environment variables are not printed unless `-vv` is set:

    + cd "/home/user/workspace" && RUST_BACKTRACE="1" CARGO_HOME="/cargo-home" "/run/user/1000/just/just-r2rLtg/configure"