Global Recipe Completion
When completions happen with a local justfile present, global recipes are ignored. All this does is allow for them to appear. A better strategy may be to consume the `-g|--global-justfile` argument and use that to inspire the autocomplete. However, since running `just <GLOBAL_RECIPE>` without the `-g` flag works, I think this may violate the best approach.